app/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Databricks integrations
databricks-vector-search>=0.22.0
databricks-sdk>=0.14.0

# LLM client
openai>=1.3.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Web scraping (for data pipeline)
beautifulsoup4>=4.12.0
requests>=2.31.0

# Spark (for Databricks notebooks)
# Note: Installed by Databricks runtime
